server/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
mysql-connector-python==8.0.28
protobuf==3.19.3
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2